Restructure functions into themes py files
Continued
